fix: added the utility function for the track events
